Rename package 'cluster' to 'service'
The package is basically a holder for utilities for working with
Dataproc. This includes jobs, workflows, etc, not just clusters.
